Adds xml-docs in stream response object + renaming a file.
